+++ playerbot/strategy/triggers/GenericTriggers.h
@@ -664,7 +664,7 @@ namespace ai
     public:
         virtual bool IsActive()
         {
-            return AI_VALUE2(Unit*, "party member without item", item) && AI_VALUE2(uint8, "item count", item);
+            return AI_VALUE2(Unit*, "party member without item", item) && AI_VALUE2(uint32, "item count", item);
         }
 
     protected:
@@ -679,7 +679,7 @@ namespace ai
     public:
         virtual bool IsActive()
         {
-            return AI_VALUE(Unit*, "party member without food") && AI_VALUE2(uint8, "item count", item);
+            return AI_VALUE(Unit*, "party member without food") && AI_VALUE2(uint32, "item count", item);
         }
     };
 
@@ -690,7 +690,7 @@ namespace ai
     public:
         virtual bool IsActive()
         {
-            return AI_VALUE(Unit*, "party member without water") && AI_VALUE2(uint8, "item count", item);
+            return AI_VALUE(Unit*, "party member without water") && AI_VALUE2(uint32, "item count", item);
         }
     };
